AI_flask_server/AI_flask_server/requirements.txt
Flask==3.0.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
Pillow==9.3.0
requests==2.28.1
